==> 2_Concentration_Factors/1_Total_output_SectorCountry-woANTUSR.ipynb <==
import pandas as pd
import numpy as np
import geopandas as gpd
import salem
import json

Export = pd.read_csv("../1_Country_Emission/Export_woANTUSR.csv")
# We want to make use of Gross_output

Export

## Question: What should we do with ROW?
#### $$\text{concentration factors} = \frac{\text{concentration caused by sector i in country j}}{\text{total output of sector i in country j}}$$
#### We only have one total output including all sectors for ROW.
#### We split the total output of ROW to different sectors according to the each sector's fraction among all sectors of other 189 countries.
#### $$\text{AGR output of ROW} = \frac{\text{AGR output of 189 countries}}{\text{total output of 189 countries}} \times \text{total output of ROW}$$

### Get rid of ANT: Netherlands Antilles (Netherlands) and USR: USSR

## Update: instead of getting rid of these two countries, using concentrations factors for the rest of the world

# Export = Export.drop(index=range(120*26,121*26),axis=0)
# Export = Export.drop(index=range(174*26,175*26),axis=0)
# # Reset column names and row indexes
# Export = Export.reset_index(drop=True)

Countries = Export['Country'].unique().tolist()

'SUD' in Countries

sector_name = Export['Industries'].unique().tolist()[:-1]

# Assign sectors in EORA to sectors in CEDS
sectors = ['AGR/RCOO','RCOO','IND','IND','IND','IND','IND','IND','IND','IND','IND','IND','ENE',
           'IND','RCOC','RCOC','RCOC','RCOC','NRTR/SHP','0','0','RCOC/RCOO','RCOC','RCOR','0','0']
agr = [0]

ene = [i for i, x in enumerate(sectors) if x == 'ENE']

ind = [i for i, x in enumerate(sectors) if x == 'IND']

nrtr = [sectors.index('NRTR/SHP')]

rcor = [i for i, x in enumerate(sectors) if x == 'RCOR']

rcoc = [i for i, x in enumerate(sectors) if x == 'RCOC']
rcoc.append(sectors.index('RCOC/RCOO'))

rcoo = [i for i, x in enumerate(sectors) if x == 'RCOO']
rcoo.append(sectors.index('RCOC/RCOO'))
rcoo.append(sectors.index('AGR/RCOO'))

shp = [sectors.index('NRTR/SHP')]

#  What belongs to IND or RCO also belongs to SLV
slv = list(np.unique(ind + rcor + rcoc + rcoo))

# Every sector will have waste.
wst = list(np.unique(slv + shp + nrtr + ene + agr))

A = ['agr','ene','ind','nrtr','rcor','rcoc','rcoo','shp','slv','wst']

sector_pos = {}
sector_pos['agr'] = [0]
sector_pos['ene'] = [i for i, x in enumerate(sectors) if x == 'ENE']

sector_pos['ind'] = [i for i, x in enumerate(sectors) if x == 'IND']

sector_pos['nrtr'] = [sectors.index('NRTR/SHP')]

sector_pos['rcor'] = [i for i, x in enumerate(sectors) if x == 'RCOR']

sector_pos['rcoc'] = [i for i, x in enumerate(sectors) if x == 'RCOC']
sector_pos['rcoc'].append(sectors.index('RCOC/RCOO'))

sector_pos['rcoo'] = [i for i, x in enumerate(sectors) if x == 'RCOO']
sector_pos['rcoo'].append(sectors.index('RCOC/RCOO'))
sector_pos['rcoo'].append(sectors.index('AGR/RCOO'))

sector_pos['shp'] = [sectors.index('NRTR/SHP')]

#  What belongs to IND or RCO also belongs to SLV
sector_pos['slv'] = list(np.unique(ind + rcor + rcoc + rcoo))

# Every sector will have waste.
sector_pos['wst'] = list(np.unique(slv + shp + nrtr + ene + agr))

sector_pos

list(np.array(sector_name)[shp])

total_output_sector = {}
for i in Countries[:-1]:
    total_output_sector[i]={}
    for j in A:
        total_output_sector[i][j]= {}
total_output_sector

for i in range(len(Countries)):
    if Countries[i] != "ROW":
        for j in A:
            temp_index = [index + 26*i for index in sector_pos[j]]
            total_output_sector[Countries[i]][j] = Export.iloc[temp_index]['Gross_output'].sum()
#     else:
#         for j in A:
#             total_output_sector[Countries[i]][j] = 0

len(total_output_sector)

total_output = {}
for i in Countries:
    total_output[i]={}

# Summation of the value insde a dictionary
def dictsum(myDict):      
    sum = 0
    for i in myDict: 
        sum = sum + myDict[i]     
    return sum

for i in Countries[:-1]:
    total_output[i] = dictsum(total_output_sector[i])
total_output['ROW'] = Export[Export['Country'] == 'ROW']['Gross_output'].sum()
total_output_sector['ROW'] = total_output['ROW']

total_output

file=open("total_output.json",'w')
json.dump(total_output,file)
file.close()

# for i in total_output_sector:
#     for j in total_output_sector[i]:
#         total_output_sector['ROW'][j] += total_output_sector[i][j]

# dominator = dictsum(total_output_sector['ROW'])
# for i in A:
#     total_output_sector['ROW'][i] = (total_output_sector['ROW'][i]/dominator) * total_output['ROW']

# # check correctness
# dictsum(total_output_sector['ROW']),total_output['ROW']

file=open("total_output_sector.json",'w')
json.dump(total_output_sector,file)
file.close()
