==> 2_Concentration_Factors/3_Concentration_factors.ipynb <==
import pandas as pd
import numpy as np
import json
import geopandas as gpd

with open("total_output_sector.json",'r') as load_f:
    total_output = json.load(load_f)
Country = list(total_output.keys())

total_output

A = ['agr','ene','ind','nrtr','rcor','rcoc','rcoo','shp','slv','wst']

# For ROW
shp = gpd.read_file(r'output' + '/' + 'output' + '_' + 'ROW' + '/' + 'ROW' + '_'+'agr' + '.' + 'shp')
for j in A[1:]:
    read_shp = gpd.read_file(r'output' + '/' + 'output' + '_' + 'ROW' + '/' + 'ROW' + '_'+j + '.' + 'shp')
    shp['PNH4'] += read_shp['PNH4']
    shp['SOA'] += read_shp['SOA']
    shp['PNO3'] += read_shp['PNO3']
    shp['PSO4'] += read_shp['PSO4']
    shp['TotalPM25'] += read_shp['TotalPM25']
shp['PNH4'] = shp['PNH4']/total_output['ROW']
shp['SOA'] = shp['SOA']/total_output['ROW']
shp['PNO3'] = shp['PNO3']/total_output['ROW']
shp['PSO4'] = shp['PSO4']/total_output['ROW']
shp['TotalPM25'] = shp['TotalPM25']/total_output['ROW']
shp.to_csv('concentration_factors' + '/' + 'ROW' + '_cf.csv' ,index=False,sep=',')

-----------------------------------
## test

shp = gpd.read_file(r'C:/Users/15311/Desktop/output_FJI/FJI_agr.shp')

shp

shp['BasePM25'].iloc[0]

for j in A[1:3]:
    read_shp = gpd.read_file(r'C:/Users/15311/Desktop' + '/' + 'output' + '_' + 'FJI' + '/' + 'FJI' + '_'+j + '.' + 'shp')
    shp['BasePM25'] += read_shp['BasePM25']

shp['BasePM25'].iloc[0]/read_shp['BasePM25'].iloc[0]

-----

# Check whether a file exit
# for i in total_output:
#     for j in A:
#         if exists(r'output' + '/' + 'output' + '_' + i + '/' + i + '_'+j + '.' + 'shp'):
#             pass
#         else:
#             print(i,j,"fail", flush=True)

for i in Country[:-1]:
    for j in A:
        if i in ['ANT', 'USR']:
            shp = gpd.read_file(r'output' + '/' + 'output' + '_' + 'ROW' + '/' + 'ROW' + '_'+j + '.' + 'shp')
#             print(i,'output' + '/' + 'output' + '_' + 'ROW' + '/' + 'ROW' + '_'+j + '.' + 'shp')
        else:
            shp = gpd.read_file(r'output' + '/' + 'output' + '_' + i + '/' + i + '_'+j + '.' + 'shp')
#             print(i,'output'+'/'+'output'+'_'+i+'/'+i+'_'+j+'.'+'shp')
        if total_output[i][j] != 0:
            shp['PNH4'] = shp['PNH4']/total_output[i][j]
            shp['SOA'] = shp['SOA']/total_output[i][j]
            shp['PNO3'] = shp['PNO3']/total_output[i][j]
            shp['PSO4'] = shp['PSO4']/total_output[i][j]
            shp['TotalPM25'] = shp['TotalPM25']/total_output[i][j]
        else:
            shp['PNH4'] = 0
            shp['SOA'] = 0
            shp['PNO3'] = 0
            shp['PSO4'] = 0
            shp['TotalPM25'] = 0          
        shp.to_csv('concentration_factors' + '/' + i + '_' + j.lower() + '_cf.csv' ,index=False,sep=',')
#         print('concentration_factors' + '/' + i + '_' + j + '_cf.csv')
